Changed per comments
